CNDB-15001: Fix testing of ann_use_synthetic_score (#1931)
Make sure that tests properly udpate `ann_use_synthetic_score` when testing different
SAI's current versions.

Also make sure that versioned vector tests run for all versions supporting vector indexes.
This includes splitting some tests that have become too long due to the additional tested
versions, to prevent timeouts.
